fix video clip path on firefox and safari
