+++ styles/LifeAtTorchbox.module.scss
@@ -3,15 +3,6 @@
     margin: 0 auto;
 }
 
-.gradientBackground {
-    // Make room for the navigation bar.
-    // Todo: choose values for this carefully when adding the page nav.
-    margin-top: -60px;
-    padding-top: 40px;
-    padding-bottom: 150px;
-    background: radial-gradient(50% 50% at 50% 50%, #2f128d 0%, #251657 100%);
-    // Break the page container
-    position: absolute;
-    left: 0;
-    right: 0;
+.indigoBackground {
+    background-color: var(--color--dark-indigo);
 }
